lesson15-1

+++ app/Http/Controllers/TasksController.php
@@ -11,11 +11,16 @@ class TasksController extends Controller
     // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+        $tasks = task::all();
 
         return view('tasks.index', [
             'tasks' => $tasks,
         ]);
+    }
+        
+        return view('welcome', $data);
     }
 
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
@@ -36,12 +41,12 @@ class TasksController extends Controller
             'content' => 'required|max:191',
         ]);
         
-        $task = new task;
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+        $request->user()->tasks()->create([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
 
-        return redirect('/');
+        return back();
     }
 
     // getでtasks/idにアクセスされた場合の「取得表示処理」
@@ -83,8 +88,11 @@ class TasksController extends Controller
     // deleteでtasks/idにアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-        $task = task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
+
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         return redirect('/');
     }
